integration of float32 into the Whisper AI (#9)
Accidentally defaulted to Int8 instead of float 32 for faster_whisper. fixed.

+++ vod_trimmer_master.py
@@ -417,13 +417,13 @@ def process_vod():
     del chunk_args  # free the view list
     del y_game      # now free the actual 1.7GB game audio buffer
 
-    # --- PHASE 2: SEMANTIC PASS (WHISPER AI) ---
+   # --- PHASE 2: SEMANTIC PASS (WHISPER AI) ---
     # Importing the WhisperModel here ensures it only loads when needed and not during GUI initialization.
     from faster_whisper import WhisperModel
     
     # NATIVE CPU ENFORCEMENT (Bypasses CUDA/PyTorch Thread Thrashing)
     print(f"\n>>> Semantic Pass (Pass 2 - {os.cpu_count() or 12} Thread AI Engine on CPU)...")
-    model = WhisperModel("base.en", device="cpu", compute_type="int8", cpu_threads=os.cpu_count() or 12)
+    model = WhisperModel("base.en", device="cpu", compute_type="float32", cpu_threads=os.cpu_count() or 12)
     raw_segments = []
     total_hype, peak_wpm, b_count = 0, 0, 0
     wpm_log = [] # Master bucket for average WPM calculation
